chore: give better names to configuration components
